pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyhydrointel"
version = "0.1.0"
description = "AI-driven hydroinformatics tools for wastewater systems."
authors = [
  { name="Mehmet Bülent Ercan", email="mehmetbercan@gmail.com" },
]
dependencies = [
  "tensorflow==2.15.0",
  "numpy>=1.23.5,<2.0",
  "pandas>=1.5,<3.0",
  "pillow>=9.5,<11",
  "matplotlib>=3.7,<4.0",
  "pyyaml>=6.0,<7.0"
]
requires-python = ">=3.9,<3.12"

[project.scripts]
pyhydrointel = "pyhydrointel.cli:main"